prelim-calendar/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
dateparser
icalendar
